Created a decorator called 'page_owner_only' which executes all sorts of sanity checks for Edit/Create Tasks (or Subtasks) and redirects to appropriate pages whenever user is unauthorized.
It can also be used for other views which require that the page owner
be the same as the current user (to guard against manually entered
urls that try to access otherwise inaccessible Tasks/SubTasks)

is_core, is_coord, and get_page_owner are now in misc/helper.py

In Task and SubTask models, I've added methods is_owner and
is_assignee. They tell you whether the user argument is an 'owner' of
the current Task or not.

URLs are now simpler - erp/task/ (for Tasks) and erp/subtask/ (for Subtasks)